+++ src/main/java/be/thomasmore/qrace/controller/StartController.java
@@ -1,4 +1,4 @@
-package be.thomasmore.qrace.controllers;
+package be.thomasmore.qrace.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
