+++ src/main/java/be/thomasmore/qrace/controller/UserController.java
@@ -1,15 +1,13 @@
-package be.thomasmore.qrace.controllers;
+package be.thomasmore.qrace.controller;
 
 import java.util.List;
 
 import be.thomasmore.qrace.model.User;
+import be.thomasmore.qrace.repository.UsersRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
 
 @Controller
 public class UserController {
