--- src/main/java/be/thomasmore/qrace/model/TimerModel.java
@@ -1,6 +1,0 @@
-package be.thomasmore.qrace.model;
-
-public class TimerModel {
-
-}
-
