--- src/main/java/be/thomasmore/qrace/repositories/GameRepository.java
@@ -1,4 +1,0 @@
-package be.thomasmore.qrace.repositories;
-
-public class GameRepository {
-}
